refactor(api): break up overlong module
move asset module under sequencer v1alpha1 module
