+++ crates/astria-sequencer/src/accounts/ics20_withdrawal.rs
@@ -6,8 +6,8 @@ use anyhow::{
 use astria_core::sequencer::v1alpha1::{
     asset,
     asset::IbcAsset,
+    transaction::action,
     Address,
-    Ics20Withdrawal,
 };
 use ibc_types::core::channel::{
     ChannelId,
@@ -29,7 +29,9 @@ use crate::{
     transaction::action_handler::ActionHandler,
 };
 
-fn withdrawal_to_unchecked_ibc_packet(withdrawal: &Ics20Withdrawal) -> IBCPacket<Unchecked> {
+fn withdrawal_to_unchecked_ibc_packet(
+    withdrawal: &action::Ics20Withdrawal,
+) -> IBCPacket<Unchecked> {
     let packet_data = withdrawal.to_fungible_token_packet_data();
     let serialized_packet_data =
         serde_json::to_vec(&packet_data).expect("can serialize FungibleTokenPacketData as JSON");
@@ -44,7 +46,7 @@ fn withdrawal_to_unchecked_ibc_packet(withdrawal: &Ics20Withdrawal) -> IBCPacket
 }
 
 #[async_trait::async_trait]
-impl ActionHandler for Ics20Withdrawal {
+impl ActionHandler for action::Ics20Withdrawal {
     #[instrument(skip(self))]
     async fn check_stateless(&self) -> Result<()> {
         ensure!(self.timeout_time() != 0, "timeout time must be non-zero",);
