+++ crates/astria-sequencer/src/authority/action.rs
@@ -5,8 +5,8 @@ use anyhow::{
 };
 use astria_core::sequencer::v1alpha1::{
     asset,
+    transaction::action::SudoAddressChangeAction,
     Address,
-    SudoAddressChangeAction,
 };
 use tracing::instrument;
