+++ crates/astria-sequencer/src/mint/action.rs
@@ -5,8 +5,8 @@ use anyhow::{
 };
 use astria_core::sequencer::v1alpha1::{
     asset,
+    transaction::action::MintAction,
     Address,
-    MintAction,
 };
 use tracing::instrument;
