+++ crates/astria-sequencer/src/service/info/mod.rs
@@ -146,12 +146,12 @@ impl Service<InfoRequest> for Info {
 
 #[cfg(test)]
 mod test {
-    use astria_core::sequencer::{
+    use astria_core::sequencer::v1alpha1::{
         asset::{
             Denom,
             DEFAULT_NATIVE_ASSET_DENOM,
         },
-        v1alpha1::Address,
+        Address,
     };
     use cnidarium::StateDelta;
     use tendermint::v0_37::abci::{
