+++ crates/astria-sequencer/src/transaction/mod.rs
@@ -11,7 +11,7 @@ use anyhow::{
 };
 use astria_core::sequencer::v1alpha1::{
     asset,
-    Action,
+    transaction::action::Action,
     Address,
     SignedTransaction,
     UnsignedTransaction,
